+++ src/errors/index.ts
@@ -10,7 +10,10 @@ export {
     notMultiplyOf8Error,
     prefixError,
     emptySettingError,
-    invalidMethodNameError
+    invalidMethodNameError,
+    simulationError,
+    emptyArrayError,
+    profilingFetchError,
 } from './instances';
 
 export {
@@ -23,5 +26,5 @@ export {
     BitError,
     MetadataError,
     SettingError,
-    EVMCallError
+    EVMCallError,
 } from './errors';